Correct usage examples in README
After the package name was updated and the exports were consolidated,
the README was never updated to reflect the new changes. This takes care
of that and slightly extends the examples so that they are more
complete. We also use `eth-query` instead of `ethjs` as that's what we
use internally at MetaMask.